refactor(storage): move workspace-root authority out of harbor-cli
Resolve the first of two duplication points flagged in #1404
(harbor-cli.ts connection reading): harbor-cli hand-rolled the same
per-platform application-data path that resolveMakaWorkspaceRoot already
computes, and had drifted (posix join on win32).

- Move resolveMakaWorkspaceRoot to @maka/storage, next to
  createConnectionStore — the path authority and the connection reader
  share one home (#1404: "shared workspace-path + connection reader /
  ConnectionStore authority").
- harbor-cli.resolveDefaultConnectionsPath now delegates to the shared
  resolver instead of re-deriving darwin/win32/linux branches. This also
  fixes a latent win32 bug: the authority uses win32 path joining, the
  old copy used posix join unconditionally.
- cli's workspace-root.ts becomes a re-export shim so the three relative
  importers (cli.ts, inspect-command.ts, run-command.ts) need no edits;
  its test moves to storage with the source.

Behavior-neutral: pure file move + path delegation. The second
duplication point (applyConnectionDefaults inline schema/ConnectionStore
routing) is a follow-up commit.
